+++ src/utils/helpers.ts
@@ -32,3 +32,29 @@ export function base64UrlDecode(base64Url: string): string {
   const decodedString = decode(paddedBase64);
   return decodedString;
 }
+
+/**
+ * Amount is represented in the smallest unit format on chain, this function converts
+ * a human readable amount format to the smallest unit format
+ * @example
+ * human readable amount format: 500
+ * on chain amount format when decimal is 8: 50000000000
+ *
+ * @param value The value in human readable format
+ * @param decimal The token decimal
+ * @returns The value is smallest units
+ */
+export const convertAmountFromHumanReadableToOnChain = (value: number, decimal: number) => value * 10 ** decimal;
+
+/**
+ * Amount is represented in the smallest unit format on chain, this function converts
+ * the smallest unit format to a human readable amount format
+ * @example
+ * human readable amount format: 500
+ * on chain amount format when decimal is 8: 50000000000
+ *
+ * @param value The value in human readable format
+ * @param decimal The token decimal
+ * @returns The value is smallest units
+ */
+export const convertAmountFromOnChainToHumanReadable = (value: number, decimal: number) => value / 10 ** decimal;